Fix SLAM

+++ src/localizers/orb_slam3/src/orb_slam3.cc
@@ -39,9 +39,9 @@ void grabImage(const sensor_msgs::ImageConstPtr &msg) {
         pose.pose.pose.position.y = Twc.at<float>(1, 3);
         pose.pose.pose.position.z = Twc.at<float>(2, 3);
 
-        tf2::Matrix3x3 tf2_rot(Twc.at<double>(0, 0), Twc.at<double>(0, 1), Twc.at<double>(0, 2),
-                               Twc.at<double>(1, 0), Twc.at<double>(1, 1), Twc.at<double>(1, 2),
-                               Twc.at<double>(2, 0), Twc.at<double>(2, 1), Twc.at<double>(2, 2));
+        tf2::Matrix3x3 tf2_rot(Twc.at<float>(0, 0), Twc.at<float>(0, 1), Twc.at<float>(0, 2),
+                               Twc.at<float>(1, 0), Twc.at<float>(1, 1), Twc.at<float>(1, 2),
+                               Twc.at<float>(2, 0), Twc.at<float>(2, 1), Twc.at<float>(2, 2));
         tf2::Quaternion tf2_quat_rot;
         tf2_rot.getRotation(tf2_quat_rot);
         pose.pose.pose.orientation = tf2::toMsg(tf2_quat_rot);
